Prepare for submoduling with libstb-secvar
This commit removes current edk2-compat-process files in preparation for submoduling with the libstb-secvar repository. Some functions will not be submoduled so they were moved to edk2-compat.c.

I did not remove ccan files, I am wondering if it would be less painful to leave them and see if we can work around having two of the same files later on. This could be a mistake come compile time. Fingers crossed!

Signed-off-by: Nick Child <nick.child@ibm.com>