Fixup secvar/backend directory for submodule compat
the old edk2-compat-process had some functions that the submoduled edk2-compat-process does not. These functions were added to `edk2-compat.c` instead. In fact, it could be argued that they should have been there in the first place. `edk2-compat.h` had to be created to allow for `edk2-compat-reset.c` to use some of the functions.

Additionally, work has begun to cleanup the Makefile paths and included directories

NOTE: At this point, skiboot is still not compiling

Signed-off-by: Nick Child <nick.child@ibm.com>